refactor: #97 - 비즈니스 로직를 useCreateLink 훅으로 분리
