+++ src/components/common/LinkList/hooks/useCreateLink.ts
@@ -1,0 +1,59 @@
+'use client'
+
+import { Dispatch, SetStateAction, useState } from 'react'
+import { UseFormSetValue } from 'react-hook-form'
+import { FetchCreateLinkProps, fetchCreateLink } from '@/services/link/link'
+import { FetchGetMetaProps, fetchGetMeta } from '@/services/meta/meta'
+import { CreateLinkFormValue } from '../LinkList'
+
+export interface UseCreateLinkReturnType {
+  isUrlCheck: boolean
+  setIsUrlCheck: Dispatch<SetStateAction<boolean>>
+  isError: boolean
+  handleGetMeta: ({ url }: FetchGetMetaProps) => Promise<void>
+  handleCreateLink: ({
+    spaceId,
+    url,
+    title,
+    tag,
+    color,
+  }: FetchCreateLinkProps) => Promise<void>
+}
+
+const useCreateLink = (
+  setValue: UseFormSetValue<CreateLinkFormValue>,
+): UseCreateLinkReturnType => {
+  const [isUrlCheck, setIsUrlCheck] = useState(false)
+  const [isError, setIsError] = useState(false)
+
+  const handleGetMeta = async ({ url }: FetchGetMetaProps) => {
+    const { data, error } = await fetchGetMeta({
+      url,
+    })
+    setValue('title', data)
+    setIsError(error)
+    if (error === false) {
+      setIsUrlCheck(true)
+    }
+  }
+
+  const handleCreateLink = async ({
+    spaceId,
+    url,
+    title,
+    tag,
+    color,
+  }: FetchCreateLinkProps) => {
+    await fetchCreateLink({
+      spaceId,
+      url,
+      title,
+      tag,
+      color,
+    })
+  }
+
+  return { isUrlCheck, setIsUrlCheck, isError, handleGetMeta, handleCreateLink }
+}
+
+export default useCreateLink
